+++ frontend/src/sky.ts
@@ -4,12 +4,14 @@ import { constants } from "./constants";
 
 export default class Sky {
     // Sky data types:
-    _skyColour: string;                 // The hex code for the sky's colour, which will be altered depending on the time of day
+    _skyColourPrimary: string; // The hex code for the sky's primary colour, which will be altered depending on the time of day
+    _skycolourSecondary: string;    // The hex code for the sky's secondary colour, which determines level of darkness
     _starDensity: number                // Value, ideally from 1 to 64, of stars in the night sky
     _starPositions: [number, number, number][]        // The x, y, and size value for each star
 
     constructor() {
-        this._skyColour = constants.ALMOST_BLACK;       // Sky is pitch black at midnight
+        this._skyColourPrimary = constants.ALMOST_BLACK;        // Sky is pitch black at midnight
+        this._skycolourSecondary = constants.GRAY_DARKEST;      // Secondary colour is also quite dark
         this._starDensity = 32;                         // Number of stars to add to the night sky
         this._starPositions = [];
         // Randomly generate new star positions
@@ -25,10 +27,26 @@ export default class Sky {
     // Takes the AM/PM cycle and current hour and sets the sky's colour based on that
     setSkyColour = (day: boolean, cycle: string, hour: number) => {
         if (day) {
-            this._skyColour = constants.YELLOW_SKY;
+            this._skyColourPrimary = constants.YELLOW_SKY;
+            this._skycolourSecondary = constants.GRAY_MEDIUM;
         } else {
-            this._skyColour = constants.ALMOST_BLACK;
+            this._skyColourPrimary = constants.ALMOST_BLACK;
+            this._skycolourSecondary = constants.GREEN_DARKEST;
         }
+    }
+
+    renderSky = (p5: P5) => {
+        // For convenience
+        const y = 0;
+        const h = constants.SCREEN_HEIGHT;
+        const c1 = p5.color(this._skyColourPrimary);
+        const c2 = p5.color(this._skycolourSecondary);
+        for (let i = y; i <= y + h; i++) {
+            let inter = p5.map(i, y, y + h, 0, 1);
+            let c = p5.lerpColor(c2, c1, inter);
+            p5.stroke(c);
+            p5.line(0, i, constants.WORLD_VIEW_WIDTH, i);
+          }
     }
 
     renderStars = (p5: P5) => {
@@ -40,12 +58,10 @@ export default class Sky {
     }
 
     render = (p5: P5, day: boolean) => {
-        p5.background(this._skyColour);
-        if (day) {
-            // Render a gradient here??
-        } else {
+        this.renderSky(p5);
+        if (!(day)) {
             this.renderStars(p5);
         }
-        p5.text(`Daytime: ${day}`, 100, 300);
+        // p5.text(`Daytime: ${day}`, 100, 300);
     }
 }